bus module refactor
moved SystemBus inside module, moved traits Bus and Access inside mod file
