--- aliusnes/src/bus/access.rs
@@ -1,4 +1,0 @@
-pub trait Access {
-    fn read(&mut self, addr: u16) -> Option<u8>;
-    fn write(&mut self, addr: u16, data: u8);
-}
