+++ aliusnes/src/bus/wram.rs
@@ -1,4 +1,4 @@
-use super::access::Access;
+use super::Access;
 use crate::{utils::int_traits::ManipulateU16, w65c816::addressing::Address};
 
 pub struct Wram {
